+++ example/lib/ui/home_screen.dart
@@ -1,0 +1,26 @@
+import 'package:flutter/material.dart';
+import 'package:router_management/router_management.dart';
+
+class HomeScreen extends StatelessWidget {
+  const HomeScreen({Key? key}) : super(key: key);
+
+  // Declare the path as constant
+  static const path = '/home/:name';
+
+  // Declare the name as constant
+  static const name = 'Home';
+
+  @override
+  Widget build(BuildContext context) {
+    final arguments = context.arguments;
+
+    return Scaffold(
+      body: Center(
+        child: Text(
+          'Welcome ${arguments.params['name']}',
+          style: const TextStyle(fontSize: 32, color: Colors.black),
+        ),
+      ),
+    );
+  }
+}