+++ example/lib/ui/not_found.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+
+class NotFound extends StatelessWidget {
+  const NotFound({Key? key}) : super(key: key);
+
+  static const path = '/404';
+
+  static const name = 'Not Found';
+
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      body: Center(
+        child: Text(
+          '404\nNot Found',
+          style: TextStyle(fontSize: 32),
+          textAlign: TextAlign.center,
+        ),
+      ),
+    );
+  }
+}