+++ lib/src/exceptions/navigation_exception.dart
@@ -1,0 +1,8 @@
+class NavigationException implements Exception {
+  final String message;
+
+  const NavigationException(this.message);
+
+  @override
+  String toString() => '$runtimeType: $message';
+}