+++ lib/src/infrastructure/extensions/page_arguments_extension.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/widgets.dart';
+
+import '../../domain/models/page_arguments.dart';
+
+/// [PageArgumentsExtension] is used to get the current [PageArguments] on the
+/// [BuildContext]
+extension PageArgumentsExtension on BuildContext {
+  /// [BuildContext.arguments] is used to get the current [PageArguments]
+  PageArguments get arguments {
+    return ModalRoute.of(this)!.settings.arguments as PageArguments;
+  }
+}