+++ CHANGELOG.md
@@ -1,7 +1,24 @@
+## [2.0.0] - Navigator 2.0
+* Break Changes: The plugin was refactored in favor of Navigator 2.0 from imperative to declarative
+* Support for Flutter Web and Deep Links
+
+## [1.3.3] - Improvements
+* Optimized the duration of custom animation to run smoothly
+
+## [1.3.2] - Improvements
+* Optimized Fade animation
+
+## [1.3.1] - Added DartDoc Comment
+
+## [1.3.0] - Breaking Changes and Bug Fixes
+* Optimized the custom transition
+* Added semantics to routes
+* Fixed fullscreenDialog with useNativeTransitions false
+
 ## [1.2.2] - Added shadow to custom transition
 * Optimized the custom transition with a shadow during the transition between screens
 
-## [1.1.1] - Breaking Changes and Bug Fixes
+## [1.1.1] -  Breaking Changes and Bug Fixes
 * Optimized the custom transition and created compatibility with newer Flutter Versions
 * Now to add a route you need to change from Router.define to Router.addRoute
 * Changed property nativeTransition to useNativeTransitions